+++ src/defaults/preferences/prefs.js
@@ -1,3 +1,0 @@
-pref("extensions.swisscoveryubbernlocations.apikey", "");
-pref("extensions.swisscoveryubbernlocations.sruurl", "https://slsp-ube.alma.exlibrisgroup.com/view/sru/41SLSP_UBE?version=1.2&operation=searchRetrieve&recordSchema=marcxml&query=alma.isbn=");
-pref("extensions.swisscoveryubbernlocations.targetField", "abstractNote");