+++ src/modules/orderNote.js
@@ -1,9 +1,0 @@
-function log (msg) {
-    console.log("[ Swisscovery UB Bern Locations ] : " + msg)
-}
-
-async function add () {
-    log("Called addOrderNote from Module")
-}
-
-export { add }